school_21_community_bot/requirements.txt
alembic==1.7.5
aiosqlite==0.20.0
async-timeout==4.0.3
asyncpg==0.30.0
greenlet==3.1.1
Mako==1.3.5
MarkupSafe==3.0.2
openpyxl==3.1.5
psycopg2==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
python-telegram-bot==21.6
pandas==2.2.3